branches, doctors, catalogs, achievements

+++ src/components/Tables/TableFive.tsx
@@ -5,12 +5,18 @@ import Link from "next/link";
 import { SetStateAction, useEffect, useState } from "react";
 
 type Catalog = {
-  id: number;
+  _id: number;
   image: string;
   title: string;
   date: string;
   document: string;
 }
+
+type CatalogsResponse = {
+  catalogs: Catalog[];
+  currentPage: number; 
+  totalPages: number;
+};
 
 const TableFive = () => { 
   const [catalogs, setCatalogs] = useState<Catalog[]>([]);
@@ -23,41 +29,45 @@ const TableFive = () => {
   
   const itemsPerPage = 15;
 
-  useEffect(() => {
-    fetchCatalogs(currentPage);
-  }, [currentPage]);
-
-  const fetchCatalogs = async (currentPage: number) => {
+  const fetchCatalogs = async (page = 1) => {
+    setLoading(true); // Mulai loading sebelum fetch data
     try {
-      const response = await fetch(`/api/catalogs?page=${currentPage}`);
-
+      const response = await fetch(`/api/catalogs?page=${page}`);
+  
       if (!response.ok) {
         throw new Error(`HTTP error! Status: ${response.status}`);
       }
-
-      const result = await response.json(); 
-      
-      setCatalogs(result.data); // Ambil hanya bagian 'data'
-      setCurrentPage(result.pagination.currentPage);
-      setTotalPages(result.pagination.totalPages);
+  
+      const result: CatalogsResponse = await response.json();
+  
+      // Update state dengan data yang valid
+      setCatalogs(result.catalogs);
+      setCurrentPage(result.currentPage || 1);
+      setTotalPages(result.totalPages || 1);
     } catch (error) {
-      console.error("Error fetching branches:", error);
+      console.error("Error fetching catalogs:", error);
     } finally {
-      setLoading(false);
+      setLoading(false); // Pastikan loading dihentikan
     }
   };
+  
 
-  const handleDeleteCatalog = async (id: string | number) => {
+  useEffect(() => {
+    fetchCatalogs(1);
+  }, []);
+
+  const handleDeleteCatalog = async (_id: string | number) => {
+    if (!selectedCatalog) return;
     try {
       setLoadingDelete(true);
-      const response = await fetch(`/api/catalogsDelete/${id}`, {
+      const response = await fetch(`/api/catalogs/${selectedCatalog._id}`, {
         method: 'DELETE',
       });
   
       if (!response.ok) {
         throw new Error(response.statusText);
       }
-      setCatalogs((prevDoctors) => prevDoctors.filter((doctor) => doctor.id !== id));
+      setCatalogs((prevCatalogs) => prevCatalogs.filter((catalog) => catalog._id !== _id));
       setSelectedCatalog(null);
       setIsOpen(false);
     } catch (error) {
@@ -70,9 +80,9 @@ const TableFive = () => {
   return (
     <div className="rounded-[10px] border border-stroke bg-white p-4 shadow-1 dark:border-dark-3 dark:bg-gray-dark dark:shadow-card sm:p-7.5">
       <div className="max-w-full overflow-x-auto">
-      {loading ? (
+      {/* {loading ? (
         <p className="text-center text-gray-500 dark:text-white mb-5 text-2xl font-semibold">Loading...</p>
-      ) : (
+      ) : ( */}
         <table className="w-full table-auto">
           <thead>
             <tr className="bg-[#F7F9FC] text-left dark:bg-dark-2">
@@ -97,7 +107,7 @@ const TableFive = () => {
             </tr>
           </thead>
           <tbody>
-            {catalogs.map((catalog, index) => (
+            {catalogs.map((catalog: any, index) => (
               <tr key={index}>
                 <td
                   className={`border-[#eee] px-4 text-center py-4 dark:border-dark-3 w-0 xl:pl-9 ${
@@ -134,7 +144,7 @@ const TableFive = () => {
                 <td
                   className={`border-[#eee] px-4 py-4 dark:border-dark-3 ${index === catalogs.length - 1 ? "border-b-0" : "border-b"}`}
                 >
-                  <Link href={`https://nmw.prahwa.net/storage/${catalog.document}`} target="blank_" className="text-dark dark:text-white">
+                  <Link href={`${catalog.document}`} target="blank_" className="text-dark dark:text-white">
                     View Document
                   </Link>
                 </td>
@@ -142,7 +152,7 @@ const TableFive = () => {
                   className={`border-[#eee] px-4 py-4 dark:border-dark-3 xl:pr-7.5 ${index === catalogs.length - 1 ? "border-b-0" : "border-b"}`}
                 >
                   <div className="flex items-center justify-end space-x-3.5">
-                    <Link href={`/catalogs/edit/${catalog.id}`} className="flex items-center justify-center">
+                    <Link href={`/catalogs/edit/${catalog._id}`} className="flex items-center justify-center">
                       <button className="hover:text-orange-400">
                         <svg xmlns="http://www.w3.org/2000/svg" width="22" height="22" viewBox="0 0 24 24"><path fill="currentColor" fillRule="evenodd" d="M21.455 5.416a.75.75 0 0 1-.096.943l-9.193 9.192a.75.75 0 0 1-.34.195l-3.829 1a.75.75 0 0 1-.915-.915l1-3.828a.8.8 0 0 1 .161-.312L17.47 2.47a.75.75 0 0 1 1.06 0l2.829 2.828a1 1 0 0 1 .096.118m-1.687.412L18 4.061l-8.518 8.518l-.625 2.393l2.393-.625z" clipRule="evenodd"/><path fill="currentColor" d="M19.641 17.16a44.4 44.4 0 0 0 .261-7.04a.4.4 0 0 1 .117-.3l.984-.984a.198.198 0 0 1 .338.127a46 46 0 0 1-.21 8.372c-.236 2.022-1.86 3.607-3.873 3.832a47.8 47.8 0 0 1-10.516 0c-2.012-.225-3.637-1.81-3.873-3.832a46 46 0 0 1 0-10.67c.236-2.022 1.86-3.607 3.873-3.832a48 48 0 0 1 7.989-.213a.2.2 0 0 1 .128.34l-.993.992a.4.4 0 0 1-.297.117a46 46 0 0 0-6.66.255a2.89 2.89 0 0 0-2.55 2.516a44.4 44.4 0 0 0 0 10.32a2.89 2.89 0 0 0 2.55 2.516c3.355.375 6.827.375 10.183 0a2.89 2.89 0 0 0 2.55-2.516"/></svg>
                       </button>
@@ -182,7 +192,7 @@ const TableFive = () => {
             ))}
           </tbody>
         </table>
-      )}
+      {/* )} */}
       {isOpen && selectedCatalog && (
         <div className="fixed top-0 left-0 w-full h-full bg-black bg-opacity-35 z-999 flex justify-center items-center z-50">
           <div className="bg-white rounded-2xl p-6 py-9 w-1/3 shadow-lg">
@@ -198,7 +208,7 @@ const TableFive = () => {
               <button className="bg-gray-200 hover:bg-gray-300 text-lg text-gray-600 py-2 px-5 rounded-lg cursor-pointer" onClick={() => setSelectedCatalog(null)}>
                 Cancel
               </button>
-              <button className="bg-red-500 hover:bg-red-600 text-lg text-white py-2 px-5 rounded-lg cursor-pointer" onClick={() => handleDeleteCatalog(selectedCatalog.id)}>
+              <button className="bg-red-500 hover:bg-red-600 text-lg text-white py-2 px-5 rounded-lg cursor-pointer" onClick={() => handleDeleteCatalog(selectedCatalog._id)}>
                 {loadingDelete ? 'Deleting...' : 'Delete'}
               </button>
             </div>
