+++ src/components/Tables/TableSix.tsx
@@ -6,12 +6,18 @@ import Link from "next/link";
 import { SetStateAction, useEffect, useState } from "react";
 
 type Achievement = {
-  id: number;
+  _id: number;
   image: string;
-  heading: string;
+  title: string;
   description: string;
   date: string;
 }
+
+type AchievementResponse = {
+  achievements: Achievement[];
+  currentPage: number;
+  totalPages: number;
+};
 
 const TableSix = () => {
   const [achievements, setAchievements] = useState<Achievement[]>([]);
@@ -24,11 +30,7 @@ const TableSix = () => {
 
   const itemsPerPage = 15;
   
-    useEffect(() => {
-      fetchAchievements(currentPage);
-    }, [currentPage]);
-  
-    const fetchAchievements = async (currentPage: number) => {
+    const fetchAchievements = async (page = 1) => {
       try {
         const response = await fetch(`/api/achievements?page=${currentPage}`);
   
@@ -36,11 +38,11 @@ const TableSix = () => {
           throw new Error(`HTTP error! Status: ${response.status}`);
         }
   
-        const result = await response.json(); 
+        const result: AchievementResponse = await response.json();
         
-        setAchievements(result.data); // Ambil hanya bagian 'data'
-        setCurrentPage(result.pagination.currentPage);
-        setTotalPages(result.pagination.totalPages);
+        setAchievements(result.achievements); // Ambil hanya bagian 'data'
+        setCurrentPage(result.currentPage);
+        setTotalPages(result.totalPages);
       } catch (error) {
         console.error("Error fetching achievements:", error);
       } finally {
@@ -48,17 +50,42 @@ const TableSix = () => {
       }
     };
 
-    const handleDeleteAchievement = async (id: string | number) => {
+    useEffect(() => {
+      fetchAchievements(1);
+    }, []);
+
+    // const handleDeleteAchievement = async (id: string | number) => {
+    //   try {
+    //     setLoadingDelete(true);
+    //     const response = await fetch(`/api/achievementsDelete/${id}`, {
+    //       method: 'DELETE',
+    //     });
+    
+    //     if (!response.ok) {
+    //       throw new Error(response.statusText);
+    //     }
+    //     setAchievements((prevAchievements) => prevAchievements.filter((achievement) => achievement.id !== id));
+    //     setSelectedAchievement(null);
+    //     setIsOpen(false);
+    //   } catch (error) {
+    //     console.error(error);
+    //   } finally {
+    //     setLoadingDelete(false);
+    //   }
+    // };
+
+    const handleDeleteAchievement = async (_id: string | number) => {
+      if (!selectedAchievement) return;
       try {
         setLoadingDelete(true);
-        const response = await fetch(`/api/achievementsDelete/${id}`, {
+        const response = await fetch(`/api/achievements/${selectedAchievement._id}`, {
           method: 'DELETE',
         });
     
         if (!response.ok) {
           throw new Error(response.statusText);
         }
-        setAchievements((prevAchievements) => prevAchievements.filter((achievement) => achievement.id !== id));
+        setAchievements((prevAchievements) => prevAchievements.filter((doctor) => doctor._id !== _id));
         setSelectedAchievement(null);
         setIsOpen(false);
       } catch (error) {
@@ -71,9 +98,9 @@ const TableSix = () => {
   return (
     <div className="rounded-[10px] border border-stroke bg-white p-4 shadow-1 dark:border-dark-3 dark:bg-gray-dark dark:shadow-card sm:p-7.5">
       <div className="max-w-full overflow-x-auto">
-        {loading ? (
+        {/* {loading ? (
           <p className="text-center text-gray-500 dark:text-white mb-5 text-2xl font-semibold">Loading...</p>
-        ) : (
+        ) : ( */}
           <table className="w-full table-auto">
             <thead>
               <tr className="bg-[#F7F9FC] text-left dark:bg-dark-2">
@@ -95,7 +122,7 @@ const TableSix = () => {
               </tr>
             </thead>
             <tbody>
-              {achievements.map((achievement, index) => (
+              {achievements.map((achievement: any, index) => (
                 <tr key={index}>
                   <td
                     className={`border-[#eee] px-4 text-center py-4 dark:border-dark-3 w-0 xl:pl-9 ${
@@ -118,7 +145,7 @@ const TableSix = () => {
                     className={`border-[#eee] px-4 py-4 dark:border-dark-3 w-170 xl:pl-0 ${index === achievements.length - 1 ? "border-b-0" : "border-b"}`}
                   >
                     <h5 className="text-dark dark:text-white">
-                      {achievement.heading}
+                      {achievement.title}
                     </h5>
                   </td>
                   <td
@@ -132,7 +159,7 @@ const TableSix = () => {
                     className={`border-[#eee] px-4 py-4 dark:border-dark-3 xl:pr-7.5 ${index === achievements.length - 1 ? "border-b-0" : "border-b"}`}
                   >
                     <div className="flex items-center justify-end space-x-3.5">
-                      <Link href={`/achievements/edit/${achievement.id}`} className="p-0 m-0 flex items-center justify-center">
+                      <Link href={`/achievements/edit/${achievement._id}`} className="p-0 m-0 flex items-center justify-center">
                         <button className="hover:text-orange-400">
                           <svg xmlns="http://www.w3.org/2000/svg" width="22" height="22" viewBox="0 0 24 24"><path fill="currentColor" fillRule="evenodd" d="M21.455 5.416a.75.75 0 0 1-.096.943l-9.193 9.192a.75.75 0 0 1-.34.195l-3.829 1a.75.75 0 0 1-.915-.915l1-3.828a.8.8 0 0 1 .161-.312L17.47 2.47a.75.75 0 0 1 1.06 0l2.829 2.828a1 1 0 0 1 .096.118m-1.687.412L18 4.061l-8.518 8.518l-.625 2.393l2.393-.625z" clipRule="evenodd"/><path fill="currentColor" d="M19.641 17.16a44.4 44.4 0 0 0 .261-7.04a.4.4 0 0 1 .117-.3l.984-.984a.198.198 0 0 1 .338.127a46 46 0 0 1-.21 8.372c-.236 2.022-1.86 3.607-3.873 3.832a47.8 47.8 0 0 1-10.516 0c-2.012-.225-3.637-1.81-3.873-3.832a46 46 0 0 1 0-10.67c.236-2.022 1.86-3.607 3.873-3.832a48 48 0 0 1 7.989-.213a.2.2 0 0 1 .128.34l-.993.992a.4.4 0 0 1-.297.117a46 46 0 0 0-6.66.255a2.89 2.89 0 0 0-2.55 2.516a44.4 44.4 0 0 0 0 10.32a2.89 2.89 0 0 0 2.55 2.516c3.355.375 6.827.375 10.183 0a2.89 2.89 0 0 0 2.55-2.516"/></svg>
                         </button>
@@ -172,7 +199,7 @@ const TableSix = () => {
               ))}
             </tbody>
           </table>
-        )}
+        {/* )} */}
         {isOpen && selectedAchievement && (
           <div className="fixed top-0 left-0 w-full h-full bg-black bg-opacity-35 z-999 flex justify-center items-center z-50">
             <div className="bg-white rounded-2xl p-6 py-9 w-1/3 shadow-lg">
@@ -182,13 +209,13 @@ const TableSix = () => {
                 </svg>
               </div>
               <p className="text-gray-600 my-5 mb-9 text-center text-2xl font-medium">
-                Are you sure you want to delete <strong>{selectedAchievement.heading}</strong>?
+                Are you sure you want to delete <strong>{selectedAchievement.title}</strong>?
               </p>
               <div className="flex justify-center gap-3">
                 <button className="bg-gray-200 hover:bg-gray-300 text-lg text-gray-600 py-2 px-5 rounded-lg cursor-pointer" onClick={() => setSelectedAchievement(null)}>
                   Cancel
                 </button>
-                <button className="bg-red-500 hover:bg-red-600 text-lg text-white py-2 px-5 rounded-lg cursor-pointer" onClick={() => handleDeleteAchievement(selectedAchievement.id)}>
+                <button className="bg-red-500 hover:bg-red-600 text-lg text-white py-2 px-5 rounded-lg cursor-pointer" onClick={() => handleDeleteAchievement(selectedAchievement._id)}>
                   {loadingDelete ? 'Deleting...' : 'Delete'}
                 </button>
               </div>
